+++ src/main/java/com/smerp/model/inventorytransactions/InventoryGoodsReceipt.java
@@ -73,7 +73,7 @@ public class InventoryGoodsReceipt extends AuditModel {
 	}
 
 	@Column(name = "freight")
-	private Integer freight;
+	private Double freight;
 
 	public Integer getId() {
 		return id;
@@ -171,11 +171,11 @@ public class InventoryGoodsReceipt extends AuditModel {
 		this.totalPayment = totalPayment;
 	}
 
-	public Integer getFreight() {
+	public Double getFreight() {
 		return freight;
 	}
 
-	public void setFreight(Integer freight) {
+	public void setFreight(Double freight) {
 		this.freight = freight;
 	}
