+++ src/main/java/com/smerp/model/inventorytransactions/InventoryGoodsTransfer.java
@@ -97,7 +97,7 @@ public class InventoryGoodsTransfer extends AuditModel {
 	}
 
 	@Column(name = "freight")
-	private Integer freight;
+	private Double freight;
 
 	public Integer getId() {
 		return id;
@@ -187,11 +187,11 @@ public class InventoryGoodsTransfer extends AuditModel {
 		this.totalPayment = totalPayment;
 	}
 
-	public Integer getFreight() {
+	public Double getFreight() {
 		return freight;
 	}
 
-	public void setFreight(Integer freight) {
+	public void setFreight(Double freight) {
 		this.freight = freight;
 	}
